A little bit of new code
Added a pause window (esc key). Also added an Enemy class- mostly
borrowed code from player class, but allows for multiple objects of an
enemy class to be created, so possibly a step toward obstacles. 3
images-Robot as generic enemy, Donut for our coin-equivalent? and a wall
as a possible jump over obstruction.